// src/app/models/vehicle.model.ts
export interface Vehicle {
  id: string;
  marca: string;
  modelo: string;
  anio: number;
  precio: number;
  kilometraje: number;
  tipoCombustible: string;
  transmision: string;
  condicion: string;
  descripcion: string;
  image: string;
  sellerName: string;
  sellerUid: string;
  sellerEmail: string;
  createdAt: string;
  updatedAt: string;
}